+++ okitweb/static/okit/js/okit_console.js
@@ -5,8 +5,9 @@
 console.debug('Loaded Console Javascript');
 
 
-const okitVersion = '0.52.0';
-const okitReleaseDate = '21st June 2023';
+const okitVersion = '0.53.0';
+const okitReleaseDate = '12th July 2023';
+// const okitReleaseDate = 'Nightly';
 
 // Validation
 const validate_error_colour = "#ff4d4d";