+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Release Notes
 
 
+## Version 0.53.0
+**Release Date**: 12th July 2023
+### Features
+1. Add platform to okit_query.py
+2. Add Auto Scaling for OCI.
+3. Add Instance Pool & Instance Configuration to PCA
+### Bug Fixes
+1. Remove agent_config from TF generated for PCA-X9 which does not support this.
+2. Fix validation error message text associated with DRG Distribution Statements.
+
+
 ## Version 0.52.0
 **Release Date**: 21st June 2023
 ### Features